clarifies "after Installation" steps
I spent a long time trying to figure out why Visual Studio Code kept reporting that my material's header files were missing even though it was being rendered correctly in Blender. 

The problem is that step 2 could be misinterpreted into believing that you should open the `.vscode` folder as a workspace instead of the root of the project. My edits clarify this by moving this extra information below as a note. 

I also did not notice the YouTube video which would have helped me understand what was happening since the iframe was broken when viewing this file in GitHub which is why I replaced it with an image + link to the video below all of the steps.